Switch to gengetopt generated command line parser
Replaces the getopt based command line parser code with auto generated
code from GNU gengetopt.

The getopt based command line parser was limited to single character
options and would have required to switch to getopt_long in the future.
Also the usage message was not up to date with the parser code.